+++ app/lib/shared/versions.dart
@@ -22,10 +22,10 @@ final RegExp runtimeVersionPattern = RegExp(r'^\d{4}\.\d{2}\.\d{2}$');
 /// when the version switch happens.
 const acceptedRuntimeVersions = <String>[
   // The current [runtimeVersion].
-  '2023.05.16',
+  '2023.05.30',
   // Fallback runtime versions.
+  '2023.05.16',
   '2023.05.10',
-  '2023.05.09', // Dart 3 analysis with beta Flutter
 ];
 
 /// Represents a combined version of the overall toolchain and processing,
@@ -54,10 +54,10 @@ bool shouldGCVersion(String version) =>
 
 // keep in-sync with SDK version in .mono_repo.yml and Dockerfile
 final String runtimeSdkVersion = '3.0.0';
-final String toolStableDartSdkVersion = '3.0.0';
-final String toolStableFlutterSdkVersion = '3.10.0';
+final String toolStableDartSdkVersion = '3.0.2';
+final String toolStableFlutterSdkVersion = '3.10.2';
 final String toolPreviewDartSdkVersion = '3.1.0-63.1.beta';
-final String toolPreviewFlutterSdkVersion = '3.11.0-0.0.pre';
+final String toolPreviewFlutterSdkVersion = '3.11.0-0.1.pre';
 
 final semanticToolStableDartSdkVersion =
     Version.parse(toolStableDartSdkVersion);