+++ CHANGELOG.md
@@ -2,6 +2,10 @@ Important changes to data models, configuration, and migrations between each
 AppEngine version, listed here to ease deployment and troubleshooting.
 
 ## Next Release (replace with git tag when deployed)
+ * Bumped runtimeVersion to `2023.05.30`.
+ * Upgraded stable Dart analysis SDK to `3.0.2`.
+ * Upgraded stable Flutter analysis SDK to `3.10.2`.
+ * Upgraded preview Flutter analysis SDK to `3.11.0-0.1.pre`.
 
 ## `20230530t100200-all`